+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/storage/services/memory/MemorySizeFinder.scala
@@ -1,12 +1,12 @@
 package uk.ac.wellcome.platform.archive.common.storage.services.memory
 
-import uk.ac.wellcome.platform.archive.common.storage.services.NewSizeFinder
+import uk.ac.wellcome.platform.archive.common.storage.services.SizeFinder
 import uk.ac.wellcome.storage.{DoesNotExistError, Identified}
 import uk.ac.wellcome.storage.store.memory.MemoryStore
 
-class NewMemorySizeFinder[Ident](
+class MemorySizeFinder[Ident](
   memoryStore: MemoryStore[Ident, Array[Byte]]
-) extends NewSizeFinder[Ident] {
+) extends SizeFinder[Ident] {
   override def get(id: Ident): ReadEither =
     memoryStore.entries.get(id) match {
       case Some(entry) => Right(Identified(id, entry.length))
