+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/storage/services/s3/S3SizeFinder.scala
@@ -2,13 +2,13 @@ package uk.ac.wellcome.platform.archive.common.storage.services.s3
 
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.AmazonS3Exception
-import uk.ac.wellcome.platform.archive.common.storage.services.NewSizeFinder
+import uk.ac.wellcome.platform.archive.common.storage.services.SizeFinder
 import uk.ac.wellcome.storage.s3.S3Errors
 import uk.ac.wellcome.storage.{DoesNotExistError, ObjectLocation, ReadError, StoreReadError}
 import uk.ac.wellcome.storage.store.RetryableReadable
 
-class NewS3SizeFinder(val maxRetries: Int = 3)(implicit s3Client: AmazonS3)
-  extends NewSizeFinder[ObjectLocation]
+class S3SizeFinder(val maxRetries: Int = 3)(implicit s3Client: AmazonS3)
+  extends SizeFinder[ObjectLocation]
     with RetryableReadable[Long] {
 
   override def retryableGetFunction(location: ObjectLocation): Long =
