+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/storage/services/memory/MemorySizeFinderTest.scala
@@ -3,10 +3,11 @@ package uk.ac.wellcome.platform.archive.common.storage.services.memory
 import java.util.UUID
 
 import uk.ac.wellcome.fixtures.TestWith
-import uk.ac.wellcome.platform.archive.common.storage.services.{NewSizeFinder, NewSizeFinderTestCases}
+import uk.ac.wellcome.platform.archive.common.storage.services.{
+  SizeFinder, SizeFinderTestCases}
 import uk.ac.wellcome.storage.store.memory.MemoryStore
 
-class NewMemorySizeFinderTest extends NewSizeFinderTestCases[UUID, MemoryStore[UUID, Array[Byte]]] {
+class MemorySizeFinderTest extends SizeFinderTestCases[UUID, MemoryStore[UUID, Array[Byte]]] {
   type StoreImpl = MemoryStore[UUID, Array[Byte]]
 
   override def withContext[R](testWith: TestWith[StoreImpl, R]): R =
@@ -14,9 +15,9 @@ class NewMemorySizeFinderTest extends NewSizeFinderTestCases[UUID, MemoryStore[U
       new MemoryStore[UUID, Array[Byte]](initialEntries = Map.empty)
     )
 
-  override def withSizeFinder[R](testWith: TestWith[NewSizeFinder[UUID], R])(
+  override def withSizeFinder[R](testWith: TestWith[SizeFinder[UUID], R])(
       implicit underlyingStore: StoreImpl): R = {
-    val sizeFinder = new NewMemorySizeFinder[UUID](
+    val sizeFinder = new MemorySizeFinder[UUID](
       memoryStore = underlyingStore
     )
